Move `Positionable`, Create `Narratable`
