+++ core/src/main/java/me/gamercoder215/socketmc/screen/Narratable.java
@@ -1,0 +1,19 @@
+package me.gamercoder215.socketmc.screen;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.io.Serializable;
+
+/**
+ * Represents a screen element that can be spoken by the narrator.
+ */
+public interface Narratable extends Serializable {
+
+    /**
+     * Gets the JSON message that represents the narration of this element. By default, this is the title of the element.
+     * @return the JSON message
+     */
+    @NotNull
+    String getNarrationMessageJSON();
+
+}
